requirements.txt
langgraph==0.5.2
langchain-core
langchain-google-genai
sentence-transformers
faiss-cpu
ipython
requests
streamlit
graphviz
